perf: only load initial resources once
